Updated TLS checks

+++ startApp.js
@@ -13,12 +13,13 @@ const SERVER_TLS_CERT = process.env.SERVER_TLS_CERT;
 const SERVER_TLS_KEY = process.env.SERVER_TLS_KEY;
 const CA_CERT = process.env.CA_CERT;
 const REQUIRE_CLIENT_CERT = process.env.REQUIRE_CLIENT_CERT || false;
+const REQUIRE_TLS = process.env.REQUIRE_TLS || false;
 
 const HOST = '0.0.0.0';
 
 //Just some stuff to get environment variables and debug, nothing interesting here...
 let debugOnOff = DEBUG?chalk.green("ON"):chalk.yellow("OFF");
-console.log("Debug is " + debugOnOff);
+console.log(chalk.white("Debug is ") + debugOnOff);
 
 if(DEBUG)
 {
@@ -37,31 +38,46 @@ if(DEBUG)
   let tls_ca = CA_CERT?chalk.green('OK'):chalk.red(undefined);
   console.log(chalk.yellow("   CA_CERT             : ") + tls_ca);
 
+  console.log(chalk.yellow("   REQUIRE_TLS         : ") + REQUIRE_TLS);
   console.log(chalk.yellow("   REQUIRE_CLIENT_CERT : ") + REQUIRE_CLIENT_CERT);
   console.log(chalk.yellow("********************************"));
 }
 
+var options = {};
+
 //If there are no references to TLS CERT/KEY then stop
-if(!SERVER_TLS_CERT || !SERVER_TLS_KEY){
-  console.log(chalk.red("The environment variable SERVER_TLS_CERT and|or SERVER_TLS_KEY was not provided. This is required for TLS. Reference the 'readme.md'."));
-  console.log("Killing process: %s", process.pid);
-  process.kill(process.pid, 'SIGTERM');
+if(REQUIRE_TLS === 'true'){
+  if(!SERVER_TLS_CERT || !SERVER_TLS_KEY){
+    console.log(chalk.red("The environment variable SERVER_TLS_CERT and|or SERVER_TLS_KEY was not provided. This is required for TLS. Reference the 'readme.md'."));
+    console.log("Killing process: %s", process.pid);
+    process.kill(process.pid, 'SIGTERM');
+  }else{
+    options.cert = fs.readFileSync(SERVER_TLS_CERT);
+    options.key = fs.readFileSync(SERVER_TLS_KEY);
+  }
+
+  //Note that the ca.crt is only needed if client certificate is requested && the provided client cert is self signed.
+  if(REQUIRE_CLIENT_CERT === 'true'){
+    options.ca = fs.readFileSync(CA_CERT);
+    options.requestCert = true;
+    options.rejectUnauthorized = true;
+  }
 }
 
-var options = {
-  cert: fs.readFileSync(SERVER_TLS_CERT),
-  key: fs.readFileSync(SERVER_TLS_KEY)
-};
+//Don't want to start both...leave as mutuially exclusive for simplicity
+let isHttps = false;
 
-//Note that the ca.crt is only needed if client certificate is requested && the provided client cert is self signed.
-if(REQUIRE_CLIENT_CERT === 'true'){
-  options.ca = fs.readFileSync(CA_CERT),
-  options.requestCert = REQUIRE_CLIENT_CERT?true:false,
-  options.rejectUnauthorized = REQUIRE_CLIENT_CERT?true:false
+if(REQUIRE_TLS === 'true'){
+  isHttps = true;
+  https.createServer(options, app).listen(HTTPS_PORT);
+}else{
+  http.createServer(app).listen(HTTP_PORT);
 }
 
-//http.createServer(app).listen(HTTP_PORT);
-https.createServer(options, app).listen(HTTPS_PORT);
+let tlsOnOff = isHttps?chalk.green('ON'):chalk.yellow('OFF');
+console.log(chalk.white("\nTLS is : ") + tlsOnOff);
 
-console.log(`Simple Node/Express API service running on => http://${HOST}:${HTTP_PORT} and https://${HOST}:${HTTPS_PORT}`);
+let serviceURI = isHttps?`https://${HOST}:${HTTPS_PORT}`:`http://${HOST}:${HTTP_PORT}`;
+
+console.log("\nService running on => %s\n", chalk.yellow(serviceURI));
